requirements.txt
paddlepaddle==2.0.2
paddlehelix
pgl
gensim==3.8.3
